Updated Workout and Program controllers to use Helper class.

+++ Controllers/WorkoutProgramController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MeFit_BE.Models;
+using MeFit_BE.Models.Domain.UserDomain;
 using MeFit_BE.Models.Domain.WorkoutDomain;
 using MeFit_BE.Models.DTO.WorkoutProgram;
 using Microsoft.AspNetCore.Authorization;
@@ -15,7 +16,7 @@ namespace MeFit_BE.Controllers
 {
     [Route("api/workout-program")]
     [ApiController]
-    //[Authorize]
+    [Authorize]
     [Produces(MediaTypeNames.Application.Json)]
     [Consumes(MediaTypeNames.Application.Json)]
     [ApiConventionType(typeof(DefaultApiConventions))]
@@ -54,20 +55,25 @@ namespace MeFit_BE.Controllers
         }
 
         /// <summary>
-        /// Method adds a new workout program to the database.
+        /// Method adds a new workout program to the database. Only a contributor can create new workout programs.
         /// </summary>
         /// <param name="programDTO">New workout program values</param>
         /// <returns>New workout program</returns>
         [HttpPost]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(404)]
+        [ProducesResponseType(403)]
         public async Task<ActionResult<WorkoutProgramReadDTO>> Post(WorkoutProgramWriteDTO programDTO)  
         {
             if (!Helper.IsContributor(HttpContext)) return Forbid();
 
-            //User user = await _context.Users.FirstOrDefaultAsync(u => u.AuthId == GetExternalUserProviderId());
-            //if (user == null) return NotFound();
+            //Find current user in the database.
+            User user = Helper.GetCurrentUser(HttpContext, _context);
+            if (user == null) return NotFound();
 
+            //Create workout program with the current user as contributor.
             WorkoutProgram domainProgram = _mapper.Map<WorkoutProgram>(programDTO);
-            //domainProgram.ContributorId = user.Id;
+            domainProgram.ContributorId = user.Id;
 
             _context.WorkoutPrograms.Add(domainProgram);
             await _context.SaveChangesAsync();
@@ -76,24 +82,27 @@ namespace MeFit_BE.Controllers
 
         /// <summary>
         /// Method updates a workout program in the database by id;
-        /// must pass in an updated workout program object
+        /// must pass in an updated workout program object. Only the contributor of the 
+        /// workout program can delete it.
         /// </summary>
         /// <param name="id">Workout program id</param>
         /// <param name="programDTO">Workout program with new values</param>
         /// <returns>Updated workout program</returns>
         [HttpPatch("{id}")]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(404)]
+        [ProducesResponseType(403)]
         public async Task<IActionResult> Put(int id, [FromBody] WorkoutProgramEditDTO programDTO)
         {
             if (!Helper.IsContributor(HttpContext)) return Forbid();
 
-            // Get WorkoutProgram
+            // Get workout program and current user from database.
             WorkoutProgram program = await GetProgramAsync(id);
-            if (program == null) return NotFound();
+            User user = Helper.GetCurrentUser(HttpContext, _context);
+            if (user == null || program == null) return NotFound();
 
-            //Ensure current contributor owns the workout program.
-            //User user = await _context.Users.FirstOrDefaultAsync(u => u.AuthId == GetExternalUserProviderId());
-            //if (user == null) return NotFound();
-            //if (program.ContributorId != user.Id) return Forbid();
+            //Ensure current contributor owns the workout program.            
+            if (program.ContributorId != user.Id) return Forbid();
 
             // Update WorkoutProgram
             if (programDTO.Name != null) program.Name = programDTO.Name;
@@ -101,32 +110,34 @@ namespace MeFit_BE.Controllers
 
             _context.WorkoutPrograms.Update(program);
             await _context.SaveChangesAsync();
-
             return Ok(program);
         }
 
         /// <summary>
-        /// Method deletes a workout program in the database by id.
+        /// Method deletes a workout program in the database by id. Only the
+        /// contributor of the workout program can delete it.
         /// </summary>
         /// <param name="id">Workout program id</param>
         [HttpDelete("{id}")]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(404)]
+        [ProducesResponseType(403)]
         public async Task<ActionResult<WorkoutProgram>> Delete(int id)
         {
             if (Helper.IsContributor(HttpContext)) return Forbid();
 
+            //Get workout program and current user form database.
             if (!ProgramExists(id))
                 return NotFound($"Program with Id: {id} was not found");
-
             WorkoutProgram program = await _context.WorkoutPrograms.FindAsync(id);
+            User user = Helper.GetCurrentUser(HttpContext, _context);
+            if (user == null) return NotFound();
 
             //Ensure current contributor owns the workout program.
-            //User user = await _context.Users.FirstOrDefaultAsync(u => u.AuthId == GetExternalUserProviderId());
-            //if (user == null) return NotFound();
-            //if (program.Id != user.Id) return Forbid();
+            if (program.Id != user.Id) return Forbid();
 
             _context.WorkoutPrograms.Remove(program);
             await _context.SaveChangesAsync();
-
             return Ok($"Deleted Program with Id: {id}");
         }
